requirements.txt
syft==0.2.8
torch==1.4.0
pyjwt==1.7.1
